+++ frontend/src/lib/components/ScheduleEditorInner.svelte
@@ -28,6 +28,7 @@
 	import FlowRetries from './flows/content/FlowRetries.svelte'
 	import WorkerTagPicker from './WorkerTagPicker.svelte'
 	import Label from './Label.svelte'
+	import DateTimeInput from './DateTimeInput.svelte'
 
 	let optionTabSelected: 'error_handler' | 'recovery_handler' | 'retries' = 'error_handler'
 
@@ -36,6 +37,7 @@
 	let edit = true
 	let schedule: string = '0 0 12 * *'
 	let timezone: string = Intl.DateTimeFormat().resolvedOptions().timeZone
+	let paused_until: string | undefined = undefined
 
 	let itemKind: 'flow' | 'script' = 'script'
 	let errorHandleritemKind: 'flow' | 'script' = 'script'
@@ -79,6 +81,8 @@
 		runnable = undefined
 		is_flow = nis_flow
 		schedule = '0 0 12 * *'
+		paused_until = undefined
+		showPauseUntil = false
 		let defaultErrorHandlerMaybe = undefined
 		let defaultRecoveryHandlerMaybe = undefined
 		if ($workspaceStore) {
@@ -248,6 +252,8 @@
 			enabled = s.enabled
 			schedule = s.schedule
 			timezone = s.timezone
+			paused_until = s.paused_until
+			showPauseUntil = paused_until !== undefined
 			summary = s.summary ?? ''
 			script_path = s.script_path ?? ''
 			await loadScript(script_path)
@@ -329,7 +335,8 @@
 					retry: retry,
 					summary: summary != '' ? summary : undefined,
 					no_flow_overlap: no_flow_overlap,
-					tag: tag
+					tag: tag,
+					paused_until: paused_until
 				}
 			})
 			sendUserToast(`Schedule ${path} updated`)
@@ -357,7 +364,8 @@
 					retry: retry,
 					summary: summary != '' ? summary : undefined,
 					no_flow_overlap: no_flow_overlap,
-					tag: tag
+					tag: tag,
+					paused_until: paused_until
 				}
 			})
 			sendUserToast(`Schedule ${path} created`)
@@ -391,6 +399,9 @@
 
 	let pathC: Path
 	let dirtyPath = false
+
+	let showPauseUntil = false
+	$: !showPauseUntil && (paused_until = undefined)
 </script>
 
 <Drawer size="900px" bind:this={drawer}>
@@ -508,6 +519,18 @@
 					<Tooltip>Schedules use CRON syntax. Seconds are mandatory.</Tooltip>
 				</svelte:fragment>
 				<CronInput disabled={!can_write} bind:schedule bind:timezone bind:validCRON />
+				<Toggle
+					options={{
+						right: 'Pause schedule until...',
+						rightTooltip:
+							'Pausing the schedule will program the next job to run as if the schedule starts at the time the pause is lifted, instead of now.'
+					}}
+					bind:checked={showPauseUntil}
+					size="xs"
+				/>
+				{#if showPauseUntil}
+					<DateTimeInput bind:value={paused_until} />
+				{/if}
 			</Section>
 			<Section label="Runnable">
 				{#if !edit}