+++ frontend/src/routes/(root)/(logged)/runs/[...path]/+page.svelte
@@ -340,37 +340,35 @@
 	async function cancelVisibleJobs() {
 		isSelectingJobsToCancel = true
 		selectedIds = jobs?.filter(isJobCancelable).map((j) => j.id) ?? []
-		if (selectedIds.length === 0 ) {
-			sendUserToast("There are no visible jobs that can be canceled", true)
+		if (selectedIds.length === 0) {
+			sendUserToast('There are no visible jobs that can be canceled', true)
 		}
 	}
 	async function cancelFilteredJobs() {
 		isCancelingFilteredJobs = true
 		fetchingFilteredJobs = true
 		const selectedFilters = {
-				workspace: $workspaceStore ?? '',
-				startedBefore: maxTs,
-				startedAfter: minTs,
-				schedulePath,
-				scriptPathExact: path === null || path === '' ? undefined : path,
-				createdBy: user === null || user === '' ? undefined : user,
-				scriptPathStart: folder === null || folder === '' ? undefined : `f/${folder}/`,
-				jobKinds,
-				success: success == 'success' ? true : success == 'failure' ? false : undefined,
-				running: success == 'running' ? true : undefined,
-				isNotSchedule: showSchedules == false ? true : undefined,
-				scheduledForBeforeNow: showFutureJobs == false ? true : undefined,
-				args:
-					argFilter && argFilter != '{}' && argFilter != '' && argError == ''
-						? argFilter
-						: undefined,
-				result:
-					resultFilter && resultFilter != '{}' && resultFilter != '' && resultError == ''
-						? resultFilter
-						: undefined,
-				allWorkspaces: allWorkspaces ? true : undefined,
-				concurrencyKey: concurrencyKey ?? undefined
-			}
+			workspace: $workspaceStore ?? '',
+			startedBefore: maxTs,
+			startedAfter: minTs,
+			schedulePath,
+			scriptPathExact: path === null || path === '' ? undefined : path,
+			createdBy: user === null || user === '' ? undefined : user,
+			scriptPathStart: folder === null || folder === '' ? undefined : `f/${folder}/`,
+			jobKinds,
+			success: success == 'success' ? true : success == 'failure' ? false : undefined,
+			running: success == 'running' ? true : undefined,
+			isNotSchedule: showSchedules == false ? true : undefined,
+			scheduledForBeforeNow: showFutureJobs == false ? true : undefined,
+			args:
+				argFilter && argFilter != '{}' && argFilter != '' && argError == '' ? argFilter : undefined,
+			result:
+				resultFilter && resultFilter != '{}' && resultFilter != '' && resultError == ''
+					? resultFilter
+					: undefined,
+			allWorkspaces: allWorkspaces ? true : undefined,
+			concurrencyKey: concurrencyKey ?? undefined
+		}
 
 		selectedFiltersString = JSON.stringify(selectedFilters, null, 4)
 		jobIdsToCancel = await JobService.listFilteredUuids(selectedFilters)
@@ -386,6 +384,11 @@
 		return j.type === 'QueuedJob' && !j.schedule_path
 	}
 
+	function jobCountString(count: number) {
+
+		return `${count} ${count == 1 ? 'job': 'jobs'}`
+	}
+
 	const warnJobLimitMsg =
 		'The exact number of concurrent job at the beginning of the time range may be incorrect as only the last 1000 jobs are taken into account: a job that was started earlier than this limit will not be taken into account'
 
@@ -394,9 +397,6 @@
 		extendedJobs !== undefined &&
 		extendedJobs.jobs.length + extendedJobs.obscured_jobs.length >= 1000
 
-	$: if (selectedIds.length === 0) {
-		isSelectingJobsToCancel = false
-	}
 </script>
 
 <JobLoader
@@ -454,7 +454,7 @@
 </ConfirmationModal>
 
 <ConfirmationModal
-	title={`Confirm cancelling the jobs visible on this page`}
+	title={`Confirm cancelling the selected jobs`}
 	confirmationText={`Cancel ${jobIdsToCancel.length} jobs`}
 	open={isCancelingVisibleJobs}
 	on:confirmed={async () => {
@@ -556,6 +556,7 @@
 			{#if graph === 'RunChart'}
 				<RunChart
 					bind:selectedIds
+					canSelect={!isSelectingJobsToCancel}
 					minTimeSet={minTs}
 					maxTimeSet={maxTs}
 					maxIsNow={maxTs == undefined}
@@ -587,13 +588,6 @@
 					{#if isSelectingJobsToCancel}
 						<div class="mt-1 p-2 h-8 flex flex-row items-center gap-1">
 							<Button
-								startIcon={{ icon: Check }}
-								size="xs"
-								color="red"
-								variant="contained"
-								on:click={cancelSelectedJobs}
-							/>
-							<Button
 								startIcon={{ icon: X }}
 								size="xs"
 								color="gray"
@@ -603,6 +597,16 @@
 									selectedIds = []
 								}}
 							/>
+							<Button
+								disabled={selectedIds.length == 0}
+								startIcon={{ icon: Check }}
+								size="xs"
+								color="red"
+								variant="contained"
+								on:click={cancelSelectedJobs}
+							>
+								Cancel {jobCountString(selectedIds.length)}
+							</Button>
 						</div>
 					{:else if !$userStore?.is_admin && !$superadmin}
 						<DropdownV2
@@ -839,6 +843,7 @@
 			{#if graph === 'RunChart'}
 				<RunChart
 					bind:selectedIds
+					canSelect={!isSelectingJobsToCancel}
 					minTimeSet={minTs}
 					maxTimeSet={maxTs}
 					maxIsNow={maxTs == undefined}
@@ -872,13 +877,6 @@
 					{#if isSelectingJobsToCancel}
 						<div class="mt-1 p-2 h-8 flex flex-row items-center gap-1">
 							<Button
-								startIcon={{ icon: Check }}
-								size="xs"
-								color="red"
-								variant="contained"
-								on:click={cancelSelectedJobs}
-							/>
-							<Button
 								startIcon={{ icon: X }}
 								size="xs"
 								color="gray"
@@ -888,6 +886,16 @@
 									selectedIds = []
 								}}
 							/>
+							<Button
+								disabled={selectedIds.length == 0}
+								startIcon={{ icon: Check }}
+								size="xs"
+								color="red"
+								variant="contained"
+								on:click={cancelSelectedJobs}
+							>
+								Cancel {jobCountString(selectedIds.length)}
+							</Button>
 						</div>
 					{:else if !$userStore?.is_admin && !$superadmin}
 						<DropdownV2